+++ src/State/ScreenProcessor.php
@@ -17,8 +17,8 @@ use App\Repository\ScreenGroupRepository;
 use App\Repository\ScreenLayoutRegionsRepository;
 use App\Repository\ScreenLayoutRepository;
 use App\Utils\IriHelperUtils;
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Uid\Ulid;
 
 class ScreenProcessor extends AbstractProcessor
 {
@@ -59,158 +59,61 @@ class ScreenProcessor extends AbstractProcessor
         // Adding relations for playlist/screen/region
         if (isset($object->regions) && isset($screen)) {
             foreach ($object->regions as $regionAndPlaylists) {
+                // Relevant region
                 $region = $this->screenLayoutRegionsRepository->findOneBy(['id' => $regionAndPlaylists['regionId']]);
 
                 if (is_null($region)) {
                     throw new InvalidArgumentException('Unknown region resource');
                 }
 
-                $newPlaylistULIDs = array_map(
-                    /**
-                     * @param array<mixed> $playlistObject
-                     *
-                     * @return Ulid
-                     */
-                    fn ($playlistObject): Ulid => Ulid::fromString($playlistObject['id']), $regionAndPlaylists['playlists']);
+                // Collection to be saved.
+                $playlistScreenRegionCollection = new ArrayCollection();
 
-                $playlistScreens = $this->playlistScreenRegionRepository->findBy([
-                    'screen' => $screen->getId(),
-                    'region' => $regionAndPlaylists['regionId'],
-                ]);
-
-                $existingPlaylistsULIDs = array_map(function ($playlistObject) {
-                    $playlist = $playlistObject->getPlaylist();
-                    if (!is_null($playlist)) {
-                        return $playlist->getId();
-                    }
-                }, $playlistScreens);
-
-                // This diff finds the playlists to be deleted
-                $deletePlaylistsULIDs = array_diff($existingPlaylistsULIDs, $newPlaylistULIDs);
-
-                // ... and deletes them.
-                foreach ($deletePlaylistsULIDs as $deletePlaylist) {
-                    $regionId = $region->getId();
-                    $screenId = $screen->getId();
-                    if (!is_null($screenId) && !is_null($regionId) && !is_null($deletePlaylist)) {
-                        $this->playlistScreenRegionRepository->deleteRelations($screenId, $regionId, $deletePlaylist);
-                    }
-                }
-
-                // This diff finds the playlists to be saved
-                $newPlaylistULIDs = array_diff($newPlaylistULIDs, $existingPlaylistsULIDs);
-
-                // ... and saves them.
-                foreach ($newPlaylistULIDs as $newPlaylist) {
-                    $playlistAndRegionToSave = new PlaylistScreenRegion();
-                    $playlist = $this->playlistRepository->findOneBy(['id' => $newPlaylist]);
+                // Looping through playlists connected to region
+                foreach ($regionAndPlaylists['playlists'] as $inputPlaylist) {
+                    // Checking if playlists exists
+                    $playlist = $this->playlistRepository->findOneBy(['id' => $inputPlaylist['id']]);
 
                     if (is_null($playlist)) {
                         throw new InvalidArgumentException('Unknown playlist resource');
                     }
 
-                    // Filter the array containing all the new playlists, to find the weight of the playlist currently
-                    // set for save
-                    $playlistWeight = array_filter($regionAndPlaylists['playlists'],
-                        /**
-                         * @param array<mixed> $playlistAndWeight
-                         *
-                         * @return bool
-                         */
-                        fn ($playlistAndWeight) => Ulid::fromString($playlistAndWeight['id']) == $playlist->getId());
-
-                    $playlistAndRegionToSave->setPlaylist($playlist);
-                    $playlistAndRegionToSave->setRegion($region);
-                    if (count($playlistWeight) > 0 && isset($playlistWeight[0]['weight'])) {
-                        $playlistAndRegionToSave->setWeight($playlistWeight[0]['weight']);
-                    } else {
-                        $playlistAndRegionToSave->setWeight(0);
-                    }
-                    $screen->addPlaylistScreenRegion($playlistAndRegionToSave);
-                }
-
-                $uneditedPlaylists = array_diff(array_diff($existingPlaylistsULIDs, $deletePlaylistsULIDs), $newPlaylistULIDs);
-
-                foreach ($existingPlaylistsULIDs as $existingPlaylist) {
-                    $region = $this->screenLayoutRegionsRepository->findOneBy(['id' => $regionAndPlaylists['regionId']]);
-
-                    if (is_null($region)) {
-                        throw new InvalidArgumentException('Unknown region resource');
-                    }
-
-                    $playlist = $this->playlistRepository->findOneBy(['id' => $existingPlaylist]);
-
-                    if (is_null($playlist)) {
-                        throw new InvalidArgumentException('Unknown playlist resource');
-                    }
-
-                    $psr = $this->playlistScreenRegionRepository->findOneBy([
-                        'screen' => $screen->getId(),
+                    // See if relation already exists
+                    $existingPlaylistScreenRegion = $this->playlistScreenRegionRepository->findOneBy([
+                        'screen' => $screen,
                         'region' => $region,
                         'playlist' => $playlist,
                     ]);
 
-                    // Filter the array containing all the new playlists, to find the weight of the playlist currently
-                    // set for save
-                    $playlistWeight = array_filter($regionAndPlaylists['playlists'],
-                        /**
-                         * @param array<mixed> $playlistAndWeight
-                         *
-                         * @return bool
-                         */
-                        fn ($playlistAndWeight) => Ulid::fromString($playlistAndWeight['id']) == $existingPlaylist);
-
-                    if (count($playlistWeight) > 0) {
-                        $psr->setWeight(reset($playlistWeight)['weight']);
+                    if (is_null($existingPlaylistScreenRegion)) {
+                        // If relation does not exist, create new PlaylistScreenRegion
+                        $newPlaylistScreenRegionRelation = new PlaylistScreenRegion();
+                        $newPlaylistScreenRegionRelation->setPlaylist($playlist);
+                        $newPlaylistScreenRegionRelation->setRegion($region);
+                        $newPlaylistScreenRegionRelation->setScreen($screen);
+                        $newPlaylistScreenRegionRelation->setWeight($inputPlaylist['weight'] ?? 0);
+                        $playlistScreenRegionCollection->add($playlistAndRegionToSave);
                     } else {
-                        $psr->setWeight(0);
+                        // Update weight, add existing relation
+                        $existingPlaylistScreenRegion->setWeight($inputPlaylist['weight'] ?? 0);
+                        $playlistScreenRegionCollection->add($existingPlaylistScreenRegion);
                     }
                 }
+                $region->setPlaylistScreenRegions($playlistScreenRegionCollection);
             }
         }
 
         // Maps ids of existing groups
         if (isset($object->groups) && isset($screen)) {
-            $existingGroups = array_map(function ($group) {
-                if (!is_null($group)) {
-                    return $group->getId();
-                }
-            }, iterator_to_array($screen->getScreenGroups()));
-
-            // Ids of groups inputted
-            $newGroupsId = array_map(
-                /**
-                 * @param string $group
-                 *
-                 * @return Ulid
-                 */
-                fn ($group): Ulid => Ulid::fromString($group), $object->groups);
-
-            // This diff finds the groups to be saved
-            $newGroups = array_diff($newGroupsId, $existingGroups);
-            // ... and saves them.
-            foreach ($newGroups as $group) {
+            $groupCollection = new ArrayCollection();
+            foreach ($object->groups as $group) {
                 $groupToSave = $this->groupRepository->findOneBy(['id' => $group]);
-
                 if (is_null($groupToSave)) {
-                    throw new InvalidArgumentException('Unknown group resource');
+                    throw new InvalidArgumentException('Unknown screen group resource');
                 }
-
-                $screen->addScreenGroup($groupToSave);
+                $groupCollection->add($groupToSave);
             }
-
-            // This diff finds the groups to be deleted
-            $deleteGroups = array_diff($existingGroups, $newGroupsId);
-            // ... and deletes them.
-            foreach ($deleteGroups as $group) {
-                $groupToDelete = $this->groupRepository->findOneBy(['id' => $group]);
-
-                if (is_null($groupToDelete)) {
-                    throw new InvalidArgumentException('Unknown group resource');
-                }
-
-                $screen->removeScreenGroup($groupToDelete);
-            }
+            $screen->setScreenGroups($groupCollection);
         }
 
         if (!empty($object->layout)) {